Switch to Community Maintained fork of bump-lavaplayer
I tried setting this up as an alternative to using red-discordbot and
ended up immediately running into #885.

There is also #1375 which advocates for moving to youtube-dl+ffmpeg as a
replacement for lavaplayer, but that would take quite a bit of work as
it is called out in the issue.

After the OG lavaplayer repo was abandoned, the community forked it and
has been maintaining it since:

https://github.com/lavalink-devs/lavaplayer

This commit moves us from using this project's forked version to the
community's fork.

I tested this with some tracks that were returning HTTP 403s and it is
now resolved (at least in my testing...)